Refactored: created new package weather and moved main activity with listener to it; moved splash to splash package
